fix : delete on relations between cart-carItem-product

+++ src/main/java/com/jeyofdev/shopping_krist/domain/product/dto/ProductDTO.java
@@ -3,7 +3,6 @@ package com.jeyofdev.shopping_krist.domain.product.dto;
 import com.jeyofdev.shopping_krist.core.enums.Color;
 import com.jeyofdev.shopping_krist.core.enums.Size;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ProductDTO(
@@ -11,8 +10,8 @@ public record ProductDTO(
         String brand,
         String name,
         String description,
-        BigDecimal price,
-        BigDecimal oldPrice,
+        Double price,
+        Double oldPrice,
         int stock,
         Color color,
         Size size
